+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-## 33.0.0 (2021-04-12)
+### 32.4.0 (2021-04-15)
+
+##### New Features
+
+* **ORCID:**  Update ORCID URL's scope (12efb3da)
+* **Signup:**  Show error message if user has already been created (19fd87f0)
+
+##### Bug Fixes
+
+* **Profile:**  Fix delete ORCID dialog not appearing (73bf8fbf)
+* **Authentication:**  Store new token (1c0d1bf5)
 
 ##### Other Changes